Factor BraveComponentExtension base class out of BaseBraveShieldsService
