+++ browser/extensions/BUILD.gn
@@ -6,6 +6,8 @@ source_set("extensions") {
     "api/brave_shields_api.h",
     "api/content_settings/brave_content_settings_store.cc",
     "api/content_settings/brave_content_settings_store.h",
+    "brave_component_extension.cc",
+    "brave_component_extension.h",
     "brave_component_extension_resource_manager.cc",
     "brave_component_extension_resource_manager.h",
     "brave_component_loader.cc",
